backend/ai-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
openai>=1.14.0
pydantic>=2.6.4
httpx>=0.27.0
Pillow>=10.2.0
python-multipart>=0.0.9
sse-starlette>=1.6.5
